+++ mongodb/rootfs/etc/mongod.d/scripts/disable-replica.js
@@ -1,0 +1,7 @@
+
+const status = rs.status();
+if (status.ok !== 0) {
+  printjson(rs.remove());
+} else {
+  print("Replica set already disabled");
+}