+++ mongodb/rootfs/etc/mongod.d/scripts/enable-replica.js
@@ -1,0 +1,8 @@
+
+const status = rs.status();
+if (status.ok === 0) {
+  printjson(rs.initiate());
+} else {
+  print("Replica set already initialized");
+}
+